Add some substance to the Address class
Rather than just having an Address class for the sake of it, it should
have some methods associated with it. There were show*, read*, to*,
from* functions for IPv4 and IPv6. The class now encapsulates that
functionality.

This change prevents the use of readIPv4, readIPv6, etc for the
Show/Read instances as they are no longer exposed. This was never ideal
anyway. For now the respective constructors are exposed and Show/Read is
derived.