Add error reporting for parsing functions
The readAddress and readSubnet class methods were replaced with
readsAddress and readsSubnet, respectively. These return ReadS to
indicate errors or ambiguity. The readAddress and readSubnet functions
were moved outside of the class and simply throw an Exception if
parsing fails.

ReadP is used to make parsing more expressive and easier to report
errors. This came with a relatively significant performance penalty for
IPv4 (from ~4us to ~8us according to the benchmark).