Use Read/Show for parsing/pretty printing
The showsAddress/readAddress and showsSubnet/readSubnet type class
methods were effectively replicating the Read/Show classes. These
methods have been removed in favor of custom Read/Show instances.

The Show documentation states that the result of 'show' should be "a
syntactically correct Haskell expression". This is not the case since
e.g. "ip = 192.168.1.1" is not a valid Haskell expression, however it
seems (ab)using Show for the purposes of "pretty printing" is
acceptable (i.e. some "standard" instances do this, such as IOError and
Handle).

This is a backwards incompatible change.