Use upTo(1) parser in readpIPv6
Add upTo and upTo1 parsers, which are the same as many and many1 but
with an upper bound on the amount of occurances. These are used to
restrict the amount of parsed fields in an IPv6 address, instead of
parsing as many as possible and failing if more than 8 are consumed.