Remove redundant parser branch in readpIPv6
The special case of compressed zeroes occurring at the end of an IPv6
address string (e.g. "1:1::") supported by the more general case of
compressed zeroes in the middle or end of the string. Generalize the
branch parsing compressed zeroes in the middle of the string to support
compressed zeroes at the end of the string, and remove the redundant
special case branch.