Add instances for HostAddress/HostAddress6
The `network` package is the de facto package for networking, and thus
this package should interoperate seamlessly with it. It is assumed the
definition of HostAddress and HostAddress does not change. As it is
a type definition as opposed to a new type the definition does not need
to be imported to reduce the amount of dependencies.

The instances for these types effectively convert the values to the
original Word32 and Word64, Word64 data types and use their respective
implementations. A proper implementation may be provided in the future.