Fix `toMask`
The `toMask` function was producing incorrect results - effectively the
inverse. For instance `(toMask 8 :: Word32)` produced 4294967040 (/24),
instead of the expected 4278190080.

Type families (requiring the TypeFamilies language extension) are used
to provide a concrete type for the Mask, given a specific `Address`,
allowing the use of `maxBound` (or `sizeOf` if a `Storable` instance for
`Word128` were available). Since a `Bounded` or `Storable` instance is
required to determine the mask's width, `Word128` from `Data.LargeWord`
is used for `IPv6`. This is a new dependency.